Feat(Application): feature BaseballGame with BaseballGame, Output Object
implement a game to match the three-digit non-duplicate  target number.When users input a three-digit number, they can check whether it is included in the target numbers and whether the digits match through balls and strikes. This will be a useful hint for inferring the target numbers.

To prevent incorrect user input, an IllegalArgumentException is raised when a number that is duplicated or has other than three digits is entered. a feature that allows the user to immediately restart the game when desired is also implemented.