Refactor(Input): refactor checkNumbersDuplicate and convertStringNumberToIntegerList method
In checkNumbersDuplicate method, "" and += has an ambiguous meaning, so modified "" to StringBuilder and += to append.

In convertStringNumberToIntegerList method, to make the meaning precise change method name from convertStringToIntegerList to convertStringNumberToIntegerList and To increase usability, instead of adding to the list, use a method of storing it in a temporary list and then assigning it as a whole.